Rename MainPage to DashboardPage and moved into appropriate folder
